requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
seaborn>=0.12.0

